[NewAccountStore] extract keytype
