+++ app/lib/modules/account/logic/key_type.dart
@@ -1,0 +1,9 @@
+enum KeyType {
+  mnemonic('mnemonic'),
+  rawSeed('rawSeed'),
+  keystore('keystore');
+
+  const KeyType(this.key);
+
+  final String key;
+}
